infra/scylladb/pyproject.toml
[tool.poetry]
name = "scylladb"
version = "0.1.0"
description = ""
authors = ["rcastillo <tepuybits@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
cassandra-driver = "^3.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
